flow: Generate the device keypair
Generate a keypair (private key and a cert, signed by the CA) to
represent the secret held on the device.

We don't do any of the CSR process, since it isn't what is being
demonstrated by this flow.  In a real application, the device key should
be kept on the device, and a CSR used to populate the certificate by the
CA server.

Signed-off-by: David Brown <david.brown@linaro.org>