flow: Give certs a distinct purpose
Match the purpose put in the OU of the given cert.  We'll have to think
about these at some point, because this is only for this flow, and there
isn't a particular reason to only use them for these purposes.  For
example, messages sent back to the device will use these same certs for
the contrary purposes.

Signed-off-by: David Brown <david.brown@linaro.org>