flow: Create a flow example for COSE_Encrypt decryption
Create an example workflow demonstrating how to decrypt the example COSE_Encrypt
packets from the cose-examples repo.  This includes a test which can be
exercised by running `cargo test` in the `tools/flow` directory.  This should be
fairly close to what is needed for decryption.  There are some missing error
checks that will result in panics instead of errors being returned.

Signed-off-by: David Brown <david.brown@linaro.org>