flow: Make recipient context optional
For the full COSE_Encrypt packet, the context was decoded as hex, because this
was useful for ensuring we got the context correct. However, this doesn't work
as the context is optional, and some packets, such as COSE_Encrypt0, do not have
a context.  The proper fix for this is to make it an `Option<Vec<u8>>` and write
an encoding module, like 'hex' that handles the field being optional.

Signed-off-by: David Brown <david.brown@linaro.org>