flow: Allow a private key to be used as well
If we detect a .pk8 file next to a certificate file, load that key as the
private key for this certificate.  We don't currently check that the private and
public keys go together, so this could potentially result in confusing errors if
the wrong private key is in the private key file.

Signed-off-by: David Brown <david.brown@linaro.org>