flow: Make the need for a private key explicit
Instead of always just looking for a parallel .pk8 file, and loading it if
present, require the caller of `Key::from_cert_file` to explicitly state whether
the private key will be needed. It will return an error if the private key is
not present and is needed. This will help clarify, in the code, which cases of a
key just need the public key from the certificate and which times need the
private key.

Signed-off-by: David Brown <david.brown@linaro.org>