sample: tfm: Update the TFM patch to add TFM_ENBALE_HEAP macro support.
The latest version of TFM 1.7 has introduced the ENABLE HEAP macro to
enable heap segments in linker files, which is by default enabled by
CMake if the build is DEBUG or Release with debug info and could not
be enabled any other way using CMake, which is incorrect and
restricts if a TF-M user wants to use heap.
So updated the existing tfm patch to introduce the
TFM_ENABLE_HEAP macro, which can be passed via
TFM_CMAKE_OPTIONS to TFM CMake.

Note:
Soon I will raise a pull request to the TF-M repo with the changes
to TFM_ENABLE_HEAP support, which can be passed via TFM_CMAKE_OPTIONS.