flow: Add a tagging module to help with COSE CBOR tags
The cose and ciborium libraries don't make it all that easy to add CBOR tags to
the COSE packets, requiring the packet to be returned as a Value, which is
wasteful.  Create a small module that is able to tag cbor packets, by calling
`encode` before writing data from a packet, and `decode` to extract the tag and
remaining packets.

Signed-off-by: David Brown <david.brown@linaro.org>